requirements.txt
# Production / CI dependencies (Phase 0-4)
yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.12.0
python-telegram-bot>=21.0
requests>=2.31.0
pyarrow>=15.0.0    # parquet I/O
